re organize tests and extraction into multiple files
